Jenkinsfile

+++ Jenkinsfile
@@ -11,7 +11,7 @@ def runtests(directory)
                 checkout([$class: 'GitSCM', branches: [[name: params.branch]], doGenerateSubmoduleConfigurations: false, extensions: [[$class: 'LocalBranch', localBranch: "**"]], submoduleCfg: [], userRemoteConfigs: [[credentialsId: '361885ba-9425-4230-950e-0af201d90547', url: 'https://git.auckland.dynabic.com/words-cloud/words-cloud-java.git']]])
                 withCredentials([usernamePassword(credentialsId: '6839cbe8-39fa-40c0-86ce-90706f0bae5d', passwordVariable: 'AppKey', usernameVariable: 'AppSid')]) {
 					sh 'mkdir -p Settings'
-                    sh 'echo "{\\"AppSid\\": \\"$AppSid\\",\\"AppKey\\": \\"$AppKey\\", \\"BaseUrl\\": \\"${testServerUrl}\\"}" > Settings/servercreds.json'
+                    sh 'echo "{\\"AppSid\\": \\"$AppSid\\",\\"AppKey\\": \\"$AppKey\\", \\"BaseUrl\\": \\"$testServerUrl\\"}" > Settings/servercreds.json'
                 }
             }
             
@@ -44,20 +44,30 @@ def runtests(directory)
 
 node('words-linux') {
 	cleanWs()
-	runtests("java-sdk")
+    if (!params.branch.contains("release")) {
+	    runtests("java-sdk")
 
-    stage('wait for publish confirmation'){
-	    timeout(time:1, unit:'DAYS') {
-		    input message:'Publish packet?'
-	    }
+        stage('wait for publish confirmation'){
+            timeout(time:1, unit:'DAYS') {
+                input message:'Publish packet?'
+            }
+        }
+
+        stage('Merge master to release'){
+            if (params.branch.contains("master")) {
+                    checkout([$class: 'GitSCM', branches: [[name: '*/release']], doGenerateSubmoduleConfigurations: false, extensions: [[$class: 'LocalBranch', localBranch: "**"]], submoduleCfg: [], userRemoteConfigs: [[credentialsId: '361885ba-9425-4230-950e-0af201d90547', url: 'https://git.auckland.dynabic.com/words-cloud/words-cloud-java.git']]])
+                    sh "git config user.email 'jenkins.aspose@gmail.com'"
+			        sh "git config user.name 'jenkins'"
+                    sh "git checkout --merge release"
+                    sh "git reset --hard origin/release"
+                    sh "git merge --no-ff --allow-unrelated-histories origin/master"
+                    sh "git diff --name-status"			
+                    sh 'git commit -am "Merged master branch to release" || exit 0'
+                    withCredentials([usernamePassword(credentialsId: '361885ba-9425-4230-950e-0af201d90547', passwordVariable: 'gitPass', usernameVariable: 'gitUsername')]) {
+                        sh "git push https://$gitUsername:$gitPass@git.auckland.dynabic.com/words-cloud/words-cloud-java.git release"
+                    }
+            }  
+        }
     }
 
-    stage('Merge master to release'){			
-	    	sh "git checkout --merge release"
-	    	sh "git reset --hard origin/release"
-	    	sh "git merge --no-ff --allow-unrelated-histories origin/master"
-	    	sh "git diff --name-status"			
-	    	sh 'git commit -am "Merged master branch to release" || exit 0'
-	    	sh "git push ${gitRepoUrl} release"
-	    }
 }
